requirements.txt
fastapi
uvicorn
python-dotenv
torch
transformers
sentencepiece
protobuf
tiktoken
scikit-learn
pandas
